requirement.txt
python-telegram-bot==20.7
requests
python-dotenv
pydub
SpeechRecognition
google-generativeai
